requirements.txt
deta==1.1.0
matplotlib==3.4.3
mysql_connector_repackaged==0.3.1
pandas<1.2.0
xlrd<2.0
Pillow==9.2.0
plotly==5.9.0
PyYAML==6.0
scikit_learn==1.1.2
streamlit==1.11.0
streamlit_authenticator==0.2.1
streamlit_option_menu==0.3.2
wordcloud==1.8.2.2